requirements.txt
opencv-python
numpy
ultralytics
Pillow
reportlab
google-cloud-vision